+++ infrastructures/symfony/Recipe/Cookbook/Enable2FA.php
@@ -63,10 +63,17 @@ class Enable2FA implements CookbookInterface
 
     protected function populateRecipe(RecipeInterface $recipe): RecipeInterface
     {
-        $recipe = $recipe->require(new Ingredient(ServerRequestInterface::class, 'request'));
-        $recipe = $recipe->require(new Ingredient('string', 'objectClass'));
-        $recipe = $recipe->require(new Ingredient('string', 'formClass'));
-        $recipe = $recipe->require(new Ingredient('array', 'formOptions'));
+        $recipe = $recipe->require(new Ingredient(requiredType: ServerRequestInterface::class, name: 'request'));
+        $recipe = $recipe->require(new Ingredient(requiredType: 'string', name: 'objectClass'));
+        $recipe = $recipe->require(new Ingredient(requiredType: 'string', name: 'formClass'));
+        $recipe = $recipe->require(
+            new Ingredient(
+                requiredType: 'array',
+                name: 'formOptions',
+                mandatory: false,
+                default: [],
+            )
+        );
 
         $recipe = $recipe->cook($this->enableTOTP, EnableTOTP::class, [], 10);