+++ src/Recipe/Cookbook/CreateObjectEndPoint.php
@@ -77,13 +77,20 @@ class CreateObjectEndPoint implements CreateObjectEndPointInterface
 
     protected function populateRecipe(RecipeInterface $recipe): RecipeInterface
     {
-        $recipe = $recipe->require(new Ingredient(ServerRequestInterface::class, 'request'));
-        $recipe = $recipe->require(new Ingredient(WriterInterface::class, 'writer'));
-        $recipe = $recipe->require(new Ingredient('string', 'route'));
-        $recipe = $recipe->require(new Ingredient('string', 'objectClass'));
-        $recipe = $recipe->require(new Ingredient('string', 'formClass'));
-        $recipe = $recipe->require(new Ingredient('array', 'formOptions'));
-        $recipe = $recipe->require(new Ingredient('string', 'template'));
+        $recipe = $recipe->require(new Ingredient(requiredType: ServerRequestInterface::class, name: 'request'));
+        $recipe = $recipe->require(new Ingredient(requiredType: WriterInterface::class, name: 'writer'));
+        $recipe = $recipe->require(new Ingredient(requiredType: 'string', name: 'route'));
+        $recipe = $recipe->require(new Ingredient(requiredType: 'string', name: 'objectClass'));
+        $recipe = $recipe->require(new Ingredient(requiredType: 'string', name: 'formClass'));
+        $recipe = $recipe->require(
+            new Ingredient(
+                requiredType: 'array',
+                name: 'formOptions',
+                mandatory: false,
+                default: [],
+            )
+        );
+        $recipe = $recipe->require(new Ingredient(requiredType: 'string', name: 'template'));
 
         $recipe = $recipe->cook($this->createObject, CreateObject::class, $this->createObjectWiths, 10);