+++ src/Recipe/Cookbook/PrepareRecoveryAccessEndPoint.php
@@ -85,15 +85,22 @@ class PrepareRecoveryAccessEndPoint implements PrepareRecoveryAccessEndPointInte
 
     protected function populateRecipe(RecipeInterface $recipe): RecipeInterface
     {
-        $recipe = $recipe->require(new Ingredient(ServerRequestInterface::class, 'request'));
-        $recipe = $recipe->require(new Ingredient(LoaderInterface::class, 'loader'));
-        $recipe = $recipe->require(new Ingredient(WriterInterface::class, 'writer'));
-        $recipe = $recipe->require(new Ingredient(AlgorithmInterface::class, 'algorithm'));
-        $recipe = $recipe->require(new Ingredient('string', 'formClass'));
-        $recipe = $recipe->require(new Ingredient('array', 'formOptions'));
-        $recipe = $recipe->require(new Ingredient('string', 'formTemplate'));
-        $recipe = $recipe->require(new Ingredient('string', 'submittedTemplate'));
-        $recipe = $recipe->require(new Ingredient('string', 'errorTemplate'));
+        $recipe = $recipe->require(new Ingredient(requiredType: ServerRequestInterface::class, name: 'request'));
+        $recipe = $recipe->require(new Ingredient(requiredType: LoaderInterface::class, name: 'loader'));
+        $recipe = $recipe->require(new Ingredient(requiredType: WriterInterface::class, name: 'writer'));
+        $recipe = $recipe->require(new Ingredient(requiredType: AlgorithmInterface::class, name: 'algorithm'));
+        $recipe = $recipe->require(new Ingredient(requiredType: 'string', name: 'formClass'));
+        $recipe = $recipe->require(
+            new Ingredient(
+                requiredType: 'array',
+                name: 'formOptions',
+                mandatory: false,
+                default: [],
+            )
+        );
+        $recipe = $recipe->require(new Ingredient(requiredType: 'string', name: 'formTemplate'));
+        $recipe = $recipe->require(new Ingredient(requiredType: 'string', name: 'submittedTemplate'));
+        $recipe = $recipe->require(new Ingredient(requiredType: 'string', name: 'errorTemplate'));
 
         $this->addToWorkplan('objectClass', EmailValue::class);
         $recipe = $recipe->cook($this->createObject, CreateObject::class, [], 10);